[fix] admin api url 변경, 토큰 재발급 api 컨트롤러 분리 (#1)
- admin api는 url에 admin이 포함되도록 수정 (여러 명이 코드 작업 시 충돌을 피하기 위함)
- 토큰 재발급 api를 admin 컨트롤러에서 token 컨트롤러로 분리
- [test_refactor] Error Response 관련 상수, 변수 BaseController로 이동
- [test_refactor] ResultActions 변수명 result로 통일, rest docs를 위한 코드는 따로 분리
