+++ src/main/java/com/example/moviebox/admin/controller/AdminController.java
@@ -2,14 +2,14 @@ package com.example.moviebox.admin.controller;
 
 import com.example.moviebox.admin.dto.AdminRequest;
 import com.example.moviebox.admin.service.AdminService;
-import com.example.moviebox.jwt.TokenDto;
+import com.example.moviebox.jwt.dto.TokenDto;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/admin")
 public class AdminController {
 	private final AdminService adminService;
 
@@ -27,10 +27,5 @@ public class AdminController {
 	@PostMapping("/login")
 	public TokenDto.Response login(@RequestBody @Valid AdminRequest adminLoginRequest) {
 		return adminService.login(adminLoginRequest.getEmail(), adminLoginRequest.getPassword());
-	}
-
-	@PostMapping("/reissue")
-	public TokenDto.Response reissue(@RequestBody @Valid TokenDto.Request tokenRequest) {	// TODO: admin에 있어도 되는가?
-		return adminService.reissue(tokenRequest);
 	}
 }
