+++ src/main/java/com/example/moviebox/admin/service/AdminService.java
@@ -3,13 +3,11 @@ package com.example.moviebox.admin.service;
 import com.example.moviebox.component.MailUtil;
 import com.example.moviebox.exception.BusinessException;
 import com.example.moviebox.jwt.*;
-import com.example.moviebox.jwt.TokenDto.Request;
-import com.example.moviebox.common.redis.RedisService;
+import com.example.moviebox.jwt.dto.TokenDto;
 import com.example.moviebox.user.domain.*;
 import java.util.regex.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,7 +22,6 @@ public class AdminService {
 	private final JwtTokenProvider jwtProvider;
 	private final MailUtil mailUtil;
 	private final UserRepository userRepository;
-	private final RedisService redisService;
 
 	@Value("${server.domain}")
 	private String serverDomain;
@@ -87,29 +84,5 @@ public class AdminService {
 		if (!passwordEncoder.matches(password, user.getPassword())) {
 			throw BusinessException.USER_NOT_FOUND_BY_PASSWORD;
 		}
-	}
-
-	@Transactional
-	public TokenDto.Response reissue(TokenDto.Request request) {
-		Long userId = validateRefreshTokenAndGetUserId(request);
-		return jwtProvider.generateAccessTokenAndRefreshToken(userId);
-	}
-
-	private Long validateRefreshTokenAndGetUserId(Request request) {
-		if (!jwtProvider.validateToken(request.getRefreshToken())) {
-			throw BusinessException.INVALID_REFRESH_TOKEN;
-		}
-		if (!jwtProvider.validateToken(request.getAccessToken())) {
-			throw BusinessException.INVALID_ACCESS_TOKEN;
-		}
-
-		Authentication authentication = jwtProvider.getAuthentication(request.getAccessToken());
-
-		String refreshToken = redisService.getTokenValues(Long.parseLong(authentication.getName()));
-		if (refreshToken == null || !refreshToken.equals(request.getRefreshToken())) {
-			throw BusinessException.EXPIRED_REFRESH_TOKEN;
-		}
-
-		return Long.parseLong(authentication.getName());
 	}
 }
