+++ src/test/java/com/example/moviebox/admin/controller/AdminControllerTest.java
@@ -15,26 +15,17 @@ import com.example.moviebox.BaseControllerTest;
 import com.example.moviebox.admin.dto.AdminRequest;
 import com.example.moviebox.admin.service.AdminService;
 import com.example.moviebox.exception.BusinessException;
-import com.example.moviebox.jwt.*;
-import java.util.Arrays;
-import java.util.List;
+import com.example.moviebox.jwt.dto.TokenDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.*;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.payload.FieldDescriptor;
 import org.springframework.test.web.servlet.ResultActions;
 
+@WebMvcTest(AdminController.class)
 @AutoConfigureRestDocs(outputDir = "build/generated-snippets")
 class AdminControllerTest extends BaseControllerTest {
-	private static final List<FieldDescriptor> ERROR_RESPONSE_FIELDS = Arrays.asList(
-		fieldWithPath("type").description("에러 코드"),
-		fieldWithPath("title").description("간략한 설명"),
-		fieldWithPath("status").description("HTTP 응답 코드"),
-		fieldWithPath("detail").description("자세한 설명"),
-		fieldWithPath("instance").description("에러 발생 근원지")
-	);
-
 	@MockBean
 	private AdminService adminService;
 
@@ -43,12 +34,16 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.register(anyString(), anyString()))
 			.willReturn(1L);
 
-		mockMvc.perform(post("/api/register")
+		ResultActions result = mockMvc.perform(post("/api/admin/register")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(new AdminRequest("example@email.com", "password"))))
+				.content(
+					objectMapper.writeValueAsString(new AdminRequest("example@email.com", "password"))
+				))
 			.andExpect(status().isOk())
-			.andExpect(content().string("1"))
-			.andDo(document("[success] register",
+			.andExpect(content().string("1"));
+
+		// docs
+		result.andDo(document("[success] register",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder()
@@ -67,13 +62,15 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.register(anyString(), anyString()))
 			.willThrow(BusinessException.EMAIL_FORMAT_INVALID);
 
-		ResultActions actions = mockMvc.perform(post("/api/register")
+		ResultActions result = mockMvc.perform(post("/api/admin/register")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsString(
 					new AdminRequest("wrong-email-format", "password"))))
 			.andExpect(status().isBadRequest());
-		checkErrorMockMvc(actions, BusinessException.EMAIL_FORMAT_INVALID);
-		actions.andDo(document("[fail] register - wrong format email",
+		checkErrorResponse(result, BusinessException.EMAIL_FORMAT_INVALID);
+
+		// docs
+		result.andDo(document("[fail] register - wrong format email",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
@@ -84,13 +81,15 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.register(anyString(), anyString()))
 			.willThrow(BusinessException.EMAIL_ALREADY_EXIST);
 
-		ResultActions actions = mockMvc.perform(post("/api/register")
+		ResultActions result = mockMvc.perform(post("/api/admin/register")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsString(
 					new AdminRequest("exist-email@email.com", "password"))))
 			.andExpect(status().isBadRequest());
-		checkErrorMockMvc(actions, BusinessException.EMAIL_ALREADY_EXIST);
-		actions.andDo(document("[fail] register - exist email",
+		checkErrorResponse(result, BusinessException.EMAIL_ALREADY_EXIST);
+
+		// docs
+		result.andDo(document("[fail] register - exist email",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
@@ -101,10 +100,12 @@ class AdminControllerTest extends BaseControllerTest {
 		willDoNothing()
 			.given(adminService).emailAuth(anyString());
 
-		mockMvc.perform(get("/api/email-auth?key=emailAuthKey"))
+		ResultActions result = mockMvc.perform(get("/api/admin/email-auth?key=emailAuthKey"))
 			.andExpect(status().isOk())
-			.andExpect(content().string("인증이 완료되었습니다."))
-			.andDo(document("[success] email auth",
+			.andExpect(content().string("인증이 완료되었습니다."));
+
+		// docs
+		result.andDo(document("[success] email auth",
 				resource(ResourceSnippetParameters.builder()
 					.summary("관리자 계정 이메일 인증")
 					.tag("admin")
@@ -120,10 +121,12 @@ class AdminControllerTest extends BaseControllerTest {
 		willThrow(BusinessException.EMAIL_AUTH_KEY_INVALID)
 			.given(adminService).emailAuth(anyString());
 
-		ResultActions actions = mockMvc.perform(get("/api/email-auth?key=wrongKey"))
+		ResultActions result = mockMvc.perform(get("/api/admin/email-auth?key=wrongKey"))
 			.andExpect(status().isBadRequest());
-		checkErrorMockMvc(actions, BusinessException.EMAIL_AUTH_KEY_INVALID);
-		actions.andDo(document("[fail] email auth - wrong key",
+		checkErrorResponse(result, BusinessException.EMAIL_AUTH_KEY_INVALID);
+
+		// docs
+		result.andDo(document("[fail] email auth - wrong key",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
@@ -139,14 +142,17 @@ class AdminControllerTest extends BaseControllerTest {
 				.build()
 			);
 
-		mockMvc.perform(post("/api/login")
+		ResultActions result = mockMvc.perform(post("/api/admin/login")
 				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(new AdminRequest("example@email.com", "password"))))
+				.content(
+					objectMapper.writeValueAsString(new AdminRequest("example@email.com", "password"))))
 			.andExpect(status().isOk())
 			.andExpect(jsonPath("$.grantType").value("Bearer"))
 			.andExpect(jsonPath("$.accessToken").value("access-token"))
-			.andExpect(jsonPath("$.refreshToken").value("refresh-token"))
-			.andDo(document("[success] login",
+			.andExpect(jsonPath("$.refreshToken").value("refresh-token"));
+
+		// docs
+		result.andDo(document("[success] login",
 				resource(ResourceSnippetParameters.builder()
 					.description("이메일과 비밀번호로 관리자 계정을 로그인할 수 있습니다.")
 					.summary("관리자 로그인")
@@ -169,13 +175,15 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.login(anyString(), anyString()))
 			.willThrow(BusinessException.USER_NOT_FOUND_BY_EMAIL);
 
-		ResultActions actions = mockMvc.perform(post("/api/login")
+		ResultActions result = mockMvc.perform(post("/api/admin/login")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsString(
 					new AdminRequest("wrong-email@email.com", "password"))))
 			.andExpect(status().isBadRequest());
-		checkErrorMockMvc(actions, BusinessException.USER_NOT_FOUND_BY_EMAIL);
-		actions.andDo(document("[fail] login - wrong email",
+		checkErrorResponse(result, BusinessException.USER_NOT_FOUND_BY_EMAIL);
+
+		// docs
+		result.andDo(document("[fail] login - wrong email",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())
@@ -187,15 +195,15 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.login(anyString(), anyString()))
 			.willThrow(BusinessException.EMAIL_NOT_VERIFIED_YET);
 
-		ResultActions actions = mockMvc.perform(post("/api/login")
+		ResultActions result = mockMvc.perform(post("/api/admin/login")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(
 					objectMapper.writeValueAsString(new AdminRequest("example@email.com", "password"))))
 			.andExpect(status().isBadRequest());
+		checkErrorResponse(result, BusinessException.EMAIL_NOT_VERIFIED_YET);
 
-		checkErrorMockMvc(actions, BusinessException.EMAIL_NOT_VERIFIED_YET);
-
-		actions.andDo(document("[fail] login - not yet email auth",
+		// docs
+		result.andDo(document("[fail] login - not yet email auth",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
@@ -206,115 +214,18 @@ class AdminControllerTest extends BaseControllerTest {
 		given(adminService.login(anyString(), anyString()))
 			.willThrow(BusinessException.USER_NOT_FOUND_BY_PASSWORD);
 
-		ResultActions actions = mockMvc.perform(post("/api/login")
+		ResultActions result = mockMvc.perform(post("/api/admin/login")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsString(
 					new AdminRequest("example@email.com", "wrong-password"))))
 			.andExpect(status().isBadRequest());
 
-		checkErrorMockMvc(actions, BusinessException.USER_NOT_FOUND_BY_PASSWORD);
+		checkErrorResponse(result, BusinessException.USER_NOT_FOUND_BY_PASSWORD);
 
-		actions.andDo(document("[fail] login - wrong password",
+		// docs
+		result.andDo(document("[fail] login - wrong password",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
 				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
-	}
-
-	@Test
-	public void testReissue() throws Exception {
-		given(adminService.reissue(any()))
-			.willReturn(TokenDto.Response.builder()
-				.grantType("Bearer")
-				.accessToken("access-token2")
-				.refreshToken("refresh-token2")
-				.build());
-
-		mockMvc.perform(post("/api/reissue")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(
-					new TokenDto.Request("access-token1", "refresh-token1")
-				)))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.grantType").value("Bearer"))
-			.andExpect(jsonPath("$.accessToken").value("access-token2"))
-			.andExpect(jsonPath("$.refreshToken").value("refresh-token2"))
-			.andDo(document("[success] reissue",
-				resource(ResourceSnippetParameters.builder()
-					.summary("JWT 토큰 재발행")
-					.tag("admin")
-					.requestFields(
-						fieldWithPath("accessToken").description("현재 Access 토큰"),
-						fieldWithPath("refreshToken").description("현재 Refresh 토큰")
-					)
-					.responseFields(
-						fieldWithPath("grantType").description("인증 타입"),
-						fieldWithPath("accessToken").description("새로 발급된 Access 토큰"),
-						fieldWithPath("refreshToken").description("새로 발급된 Refresh 토큰")
-					)
-					.build())
-			));
-	}
-
-	@Test
-	public void testReissueByWrongRefreshToken() throws Exception {
-		given(adminService.reissue(any()))
-			.willThrow(BusinessException.INVALID_REFRESH_TOKEN);
-
-		ResultActions actions = mockMvc.perform(post("/api/reissue")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(
-					new TokenDto.Request("access-token1", "wrong-refresh-token1")
-				)))
-			.andExpect(status().isUnauthorized());
-		checkErrorMockMvc(actions, BusinessException.INVALID_REFRESH_TOKEN);
-		actions.andDo(document("[fail] reissue - wrong refresh token",
-				preprocessRequest(prettyPrint()),
-				preprocessResponse(prettyPrint()),
-				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
-	}
-
-	@Test
-	public void testReissueByWrongAccessToken() throws Exception {
-		given(adminService.reissue(any()))
-			.willThrow(BusinessException.INVALID_ACCESS_TOKEN);
-
-		ResultActions actions = mockMvc.perform(post("/api/reissue")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(
-					new TokenDto.Request("wrong-access-token1", "refresh-token1")
-				)))
-			.andExpect(status().isUnauthorized());
-		checkErrorMockMvc(actions, BusinessException.INVALID_ACCESS_TOKEN);
-		actions.andDo(
-			document("[fail] reissue - wrong access token",
-				preprocessRequest(prettyPrint()),
-				preprocessResponse(prettyPrint()),
-				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
-	}
-
-	@Test
-	public void testReissueByExpiredRefreshToken() throws Exception {
-		given(adminService.reissue(any()))
-			.willThrow(BusinessException.EXPIRED_REFRESH_TOKEN);
-
-		ResultActions actions = mockMvc.perform(post("/api/reissue")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(
-					new TokenDto.Request("access-token1", "expired-refresh-token1")
-				)))
-			.andExpect(status().isUnauthorized());
-		checkErrorMockMvc(actions, BusinessException.EXPIRED_REFRESH_TOKEN);
-		actions.andDo(
-			document("[fail] reissue - expired refresh token",
-				preprocessRequest(prettyPrint()),
-				preprocessResponse(prettyPrint()),
-				resource(ResourceSnippetParameters.builder().tag("admin").responseFields(ERROR_RESPONSE_FIELDS).build())));
-	}
-
-	private void checkErrorMockMvc(ResultActions resultActions, BusinessException ex) throws Exception {
-		resultActions.andExpect(jsonPath("$.type").value(ex.getErrorCode().getErrorType()))
-			.andExpect(jsonPath("$.title").value(ex.getErrorCode().getDescription()))
-			.andExpect(jsonPath("$.status").value(ex.getHttpStatus().value()))
-			.andExpect(jsonPath("$.detail").value(ex.getMessage()));
 	}
 }
